+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Infrastructure/ConversationAnalysisTestBase.cs
@@ -11,12 +11,12 @@ namespace Azure.AI.Language.Conversations.Tests
     /// </summary>
     /// <typeparam name="TClient">The type of client being tested.</typeparam>
     [ClientTestFixture(
-        ConversationsClientOptions.ServiceVersion.V2023_04_01,
-        ConversationsClientOptions.ServiceVersion.V2022_05_01)]
+        ConversationAnalysisClientOptions.ServiceVersion.V2023_04_01,
+        ConversationAnalysisClientOptions.ServiceVersion.V2022_05_01)]
     [IgnoreServiceError(429, "429")]
     public abstract class ConversationAnalysisTestBase<TClient> : RecordedTestBase<ConversationAnalysisTestEnvironment> where TClient : class
     {
-        protected ConversationAnalysisTestBase(bool isAsync, ConversationsClientOptions.ServiceVersion serviceVersion, RecordedTestMode? mode)
+        protected ConversationAnalysisTestBase(bool isAsync, ConversationAnalysisClientOptions.ServiceVersion serviceVersion, RecordedTestMode? mode)
             : base(isAsync, mode)
         {
             // TODO: Compare bodies again when https://github.com/Azure/azure-sdk-for-net/issues/22219 is resolved.
@@ -34,7 +34,7 @@ namespace Azure.AI.Language.Conversations.Tests
         /// <summary>
         /// Gets the service version used for this instance of the test fixture.
         /// </summary>
-        protected ConversationsClientOptions.ServiceVersion ServiceVersion { get; }
+        protected ConversationAnalysisClientOptions.ServiceVersion ServiceVersion { get; }
 
         /// <summary>
         /// Creates the <see cref="Client"/> once tests begin.
@@ -43,7 +43,7 @@ namespace Azure.AI.Language.Conversations.Tests
         {
             await base.StartTestRecordingAsync();
 
-            ConversationsClientOptions options = new ConversationsClientOptions(ServiceVersion);
+            ConversationAnalysisClientOptions options = new ConversationAnalysisClientOptions(ServiceVersion);
             Client = CreateClient<TClient>(
                 TestEnvironment.Endpoint,
                 new AzureKeyCredential(TestEnvironment.ApiKey),