+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Samples/Sample2_AnalyzeConversation_OrchestrationPrediction.cs
@@ -10,13 +10,13 @@ using NUnit.Framework;
 
 namespace Azure.AI.Language.Conversations.Tests.Samples
 {
-    public partial class ConversationsClientSamples
+    public partial class ConversationAnalysisClientSamples
     {
         [SyncOnly]
         [RecordedTest]
         public void AnalyzeConversationOrchestrationPredictionQuestionAnswering()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
 
             #region Snippet:ConversationAnalysis_AnalyzeConversationOrchestrationPrediction
             string projectName = "DomainOrchestrator";
@@ -48,7 +48,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Kind = "Conversation",
             };
 
-            Response response = client.AnalyzeConversations(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = client.AnalyzeConversation(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;
@@ -79,7 +79,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         [RecordedTest]
         public void AnalyzeConversationOrchestrationPredictionConversation()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             var data = new
             {
                 AnalysisInput = new
@@ -102,7 +102,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Kind = "Conversation",
             };
 
-            Response response = client.AnalyzeConversations(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = client.AnalyzeConversation(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;
@@ -134,7 +134,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/29136")]
         public void AnalyzeConversationOrchestrationPredictionLuis()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             var data = new
             {
                 AnalysisInput = new
@@ -157,7 +157,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Kind = "Conversation",
             };
 
-            Response response = client.AnalyzeConversations(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = client.AnalyzeConversation(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;
@@ -181,7 +181,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         [RecordedTest]
         public async Task AnalyzeConversationOrchestrationPredictionQuestionAnsweringAsync()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
 
             string projectName = TestEnvironment.OrchestrationProjectName;
             string deploymentName = TestEnvironment.OrchestrationDeploymentName;
@@ -209,7 +209,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
             };
 
             #region Snippet:ConversationAnalysis_AnalyzeConversationOrchestrationPredictionAsync
-            Response response = await client.AnalyzeConversationsAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = await client.AnalyzeConversationAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;
@@ -238,7 +238,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         [RecordedTest]
         public async Task AnalyzeConversationOrchestrationPredictionConversationAsync()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             var data = new
             {
                 AnalysisInput = new
@@ -261,7 +261,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Kind = "Conversation",
             };
 
-            Response response = await client.AnalyzeConversationsAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = await client.AnalyzeConversationAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;
@@ -318,7 +318,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
         [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/29136")]
         public async Task AnalyzeConversationOrchestrationPredictionLuisAsync()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             var data = new
             {
                 AnalysisInput = new
@@ -341,7 +341,7 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                 Kind = "Conversation",
             };
 
-            Response response = await client.AnalyzeConversationsAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
+            Response response = await client.AnalyzeConversationAsync(RequestContent.Create(data, JsonPropertyNames.CamelCase));
 
             dynamic conversationalTaskResult = response.Content.ToDynamicFromJson(JsonPropertyNames.CamelCase);
             dynamic orchestrationPrediction = conversationalTaskResult.Result.Prediction;