+++ sdk/cognitivelanguage/Azure.AI.Language.Conversations/tests/Samples/Sample6_AnalyzeConversation_ConversationSummarization.cs
@@ -3,6 +3,8 @@
 
 using System;
 using System.Collections.Generic;
+using System.Data;
+using System.Linq;
 using System.Threading.Tasks;
 using Azure.AI.Language.Conversations.Models;
 using Azure.Core.TestFramework;
@@ -10,60 +12,69 @@ using NUnit.Framework;
 
 namespace Azure.AI.Language.Conversations.Tests.Samples
 {
-    public partial class ConversationsClientSamples
+    public partial class ConversationAnalysisClientSamples
     {
         [SyncOnly]
         [RecordedTest]
-        [ServiceVersion(Min = ConversationsClientOptions.ServiceVersion.V2023_04_01)]
+        [ServiceVersion(Min = ConversationAnalysisClientOptions.ServiceVersion.V2023_04_01)]
         public void AnalyzeConversation_ConversationSummarization()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             List<string> aspects = new();
 
             #region Snippet:AnalyzeConversation_ConversationSummarization
-            var data = new AnalyzeConversationJobsInput(
-                new MultiLanguageConversationAnalysisInput(
+            var data = new AnalyzeConversationOperationInput(
+                new MultiLanguageConversationInput(
                     new List<ConversationInput>
                     {
                         new TextConversation("1", "en", new List<TextConversationItem>()
                         {
-                            AILanguageConversationsModelFactory.TextConversationItem("1", "Agent_1", "Hello, how can I help you?", role: ParticipantRole.Agent),
-                            AILanguageConversationsModelFactory.TextConversationItem("2", "Customer_1", "How to upgrade Office? I am getting error messages the whole day.", role: ParticipantRole.Customer),
-                            AILanguageConversationsModelFactory.TextConversationItem("3", "Agent_1", "Press the upgrade button please. Then sign in and follow the instructions.", role: ParticipantRole.Agent)
+                            new TextConversationItem("1", "Agent_1", "Hello, how can I help you?")
+                            {
+                                Role = ParticipantRole.Agent
+                            },
+                            new TextConversationItem("2", "Customer_1", "How to upgrade Office? I am getting error messages the whole day.")
+                            {
+                                Role = ParticipantRole.Customer
+                            },
+                            new TextConversationItem("3", "Agent_1", "Press the upgrade button please. Then sign in and follow the instructions.")
+                            {
+                                Role = ParticipantRole.Agent
+                            }
                         })
                     }),
-                    new List<AnalyzeConversationJobTask>
+                    new List<AnalyzeConversationOperationAction>
                     {
-                        new AnalyzeConversationSummarizationTask()
+                        new SummarizationOperationAction()
                         {
-                            Parameters = new ConversationSummarizationTaskContent(new List<SummaryAspect>
+                            ActionContent = new ConversationSummarizationActionContent(new List<SummaryAspect>
                             {
                                 SummaryAspect.Issue,
                             }),
-                            TaskName = "Issue task",
+                            Name = "Issue task",
                         },
-                        new AnalyzeConversationSummarizationTask()
+                        new SummarizationOperationAction()
                         {
-                            Parameters = new ConversationSummarizationTaskContent(new List<SummaryAspect>
+                            ActionContent = new ConversationSummarizationActionContent(new List<SummaryAspect>
                             {
                                 SummaryAspect.Resolution,
                             }),
-                            TaskName = "Resolution task",
+                            Name = "Resolution task",
                         }
                     });
 
-            Response<AnalyzeConversationJobState> analyzeConversationOperation = client.AnalyzeConversationsOperation(data);
+            Response<AnalyzeConversationOperationState> analyzeConversationOperation = client.AnalyzeConversationOperation(data);
 
-            dynamic jobResults = analyzeConversationOperation.Value;
-            foreach (dynamic task in jobResults.Tasks.Items)
+            AnalyzeConversationOperationState jobResults = analyzeConversationOperation.Value;
+            foreach (SummarizationOperationResult task in jobResults.Actions.Items.Cast<SummarizationOperationResult>())
             {
-                Console.WriteLine($"Task name: {task.TaskName}");
-                dynamic results = task.Results;
-                foreach (dynamic conversation in results.Conversations)
+                Console.WriteLine($"Task name: {task.Name}");
+                SummaryResult results = task.Results;
+                foreach (ConversationsSummaryResult conversation in results.Conversations)
                 {
                     Console.WriteLine($"Conversation: #{conversation.Id}");
                     Console.WriteLine("Summaries:");
-                    foreach (dynamic summary in conversation.Summaries)
+                    foreach (SummaryResultItem summary in conversation.Summaries)
                     {
                         Console.WriteLine($"Text: {summary.Text}");
                         Console.WriteLine($"Aspect: {summary.Aspect}");
@@ -71,10 +82,10 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                         aspects.Add(summary.Aspect);
 #endif
                     }
-                    if (results.Warnings != null)
+                    if (conversation.Warnings != null && conversation.Warnings.Any())
                     {
                         Console.WriteLine("Warnings:");
-                        foreach (dynamic warning in conversation.Warnings)
+                        foreach (InputWarning warning in conversation.Warnings)
                         {
                             Console.WriteLine($"Code: {warning.Code}");
                             Console.WriteLine($"Message: {warning.Message}");
@@ -82,10 +93,10 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                     }
                     Console.WriteLine();
                 }
-                if (results.Errors != null)
+                if (results.Errors != null && results.Errors.Any())
                 {
                     Console.WriteLine("Errors:");
-                    foreach (dynamic error in results.Errors)
+                    foreach (DocumentError error in results.Errors)
                     {
                         Console.WriteLine($"Error: {error}");
                     }
@@ -99,57 +110,57 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
 
         [AsyncOnly]
         [RecordedTest]
-        [ServiceVersion(Min = ConversationsClientOptions.ServiceVersion.V2023_04_01)]
+        [ServiceVersion(Min = ConversationAnalysisClientOptions.ServiceVersion.V2023_04_01)]
         public async Task AnalyzeConversationAsync_ConversationSummarization()
         {
-            ConversationsClient client = Client;
+            ConversationAnalysisClient client = Client;
             List<string> aspects = new();
 
-            var data = new AnalyzeConversationJobsInput(
-                new MultiLanguageConversationAnalysisInput(
+            #region Snippet:AnalyzeConversationAsync_ConversationSummarization
+
+            var data = new AnalyzeConversationOperationInput(
+                new MultiLanguageConversationInput(
                     new List<ConversationInput>
                     {
                         new TextConversation("1", "en", new List<TextConversationItem>()
                         {
-                            AILanguageConversationsModelFactory.TextConversationItem("1", "Agent", "Hello, how can I help you?", role: ParticipantRole.Agent),
-                            AILanguageConversationsModelFactory.TextConversationItem("2", "Customer", "How to upgrade Office? I am getting error messages the whole day.", role: ParticipantRole.Customer),
-                            AILanguageConversationsModelFactory.TextConversationItem("3", "Agent", "Press the upgrade button please. Then sign in and follow the instructions.", role: ParticipantRole.Agent)
+                            new TextConversationItem("1", "Agent", "Hello, how can I help you?"),
+                            new TextConversationItem("2", "Customer", "How to upgrade Office? I am getting error messages the whole day."),
+                            new TextConversationItem("3", "Agent", "Press the upgrade button please. Then sign in and follow the instructions.")
                         })
                     }),
-                    new List<AnalyzeConversationJobTask>
+                    new List<AnalyzeConversationOperationAction>
                     {
-                        new AnalyzeConversationSummarizationTask()
+                        new SummarizationOperationAction()
                         {
-                            Parameters = new ConversationSummarizationTaskContent(new List<SummaryAspect>
+                            ActionContent = new ConversationSummarizationActionContent(new List<SummaryAspect>
                             {
                                 SummaryAspect.Issue,
                             }),
-                            TaskName = "Issue task",
+                            Name = "Issue task",
                         },
-                        new AnalyzeConversationSummarizationTask()
+                        new SummarizationOperationAction()
                         {
-                            Parameters = new ConversationSummarizationTaskContent(new List<SummaryAspect>
+                            ActionContent = new ConversationSummarizationActionContent(new List<SummaryAspect>
                             {
                                 SummaryAspect.Resolution,
                             }),
-                            TaskName = "Resolution task",
+                            Name = "Resolution task",
                         }
                     });
 
-            #region Snippet:AnalyzeConversationAsync_ConversationSummarization
-            var analyzeConversationOperation = await client.AnalyzeConversationsOperationAsync(data);
-            #endregion
+            Response<AnalyzeConversationOperationState> analyzeConversationOperation = await client.AnalyzeConversationOperationAsync(data);
 
-            AnalyzeConversationJobState jobResults = analyzeConversationOperation.Value;
-            foreach (dynamic task in jobResults.Tasks.Items)
+            AnalyzeConversationOperationState jobResults = analyzeConversationOperation.Value;
+            foreach (SummarizationOperationResult task in jobResults.Actions.Items.Cast<SummarizationOperationResult>())
             {
-                Console.WriteLine($"Task name: {task.TaskName}");
-                dynamic results = task.Results;
-                foreach (dynamic conversation in results.Conversations)
+                Console.WriteLine($"Task name: {task.Name}");
+                SummaryResult results = task.Results;
+                foreach (ConversationsSummaryResult conversation in results.Conversations)
                 {
                     Console.WriteLine($"Conversation: #{conversation.Id}");
                     Console.WriteLine("Summaries:");
-                    foreach (dynamic summary in conversation.Summaries)
+                    foreach (SummaryResultItem summary in conversation.Summaries)
                     {
                         Console.WriteLine($"Text: {summary.Text}");
                         Console.WriteLine($"Aspect: {summary.Aspect}");
@@ -157,9 +168,27 @@ namespace Azure.AI.Language.Conversations.Tests.Samples
                         aspects.Add(summary.Aspect);
 #endif
                     }
+                    if (conversation.Warnings != null && conversation.Warnings.Any())
+                    {
+                        Console.WriteLine("Warnings:");
+                        foreach (InputWarning warning in conversation.Warnings)
+                        {
+                            Console.WriteLine($"Code: {warning.Code}");
+                            Console.WriteLine($"Message: {warning.Message}");
+                        }
+                    }
                     Console.WriteLine();
                 }
+                if (results.Errors != null && results.Errors.Any())
+                {
+                    Console.WriteLine("Errors:");
+                    foreach (DocumentError error in results.Errors)
+                    {
+                        Console.WriteLine($"Error: {error}");
+                    }
+                }
             }
+            #endregion
 
             Assert.That(aspects, Contains.Item("issue").And.Contains("resolution"));
             Assert.That(analyzeConversationOperation.GetRawResponse().Status, Is.EqualTo(200));